94.980% - test: make metacov work with xdist
I was using pytest.__file__ as a directory, but in 5.3.2, pytest changed
to be a package, so that wasn't the site-packages directory anymore, and
our .pth file was written someplace useless.

Now we don't rely on the structure of pytest (why did we ever?), and it
works again.

https://nedbat.github.io/coverage-reports/reports/20211008_c163cdc449/htmlcov
c163cdc449: master