95.767% - fix: find_writable_pth_directory must expect failure from open()
Fix find_writable_pth_directory() to expect the OSError from the open()
call.  In general, this is what's going to happen if the file is not
writable.  The current logic could only be triggered e.g. if the file
was writable but the underlying filesystem run out of space.

https://nedbat.github.io/coverage-reports/reports/20211012_ce0e38e131/htmlcov
ce0e38e131: master