95.438% - fix: ensure absolute paths are relative when combined #1752
If two data files are combined, one with absolute paths, and one with relative,
with relative_files=True in effect, the results depended on the order of
combining.

If the absolute files were seen first, they were added as absolute paths.  If
the relative files were seen first, a mapping rule was generated that would then
remap the absolute paths when they were seen.

This fix ensures that absolute paths are remapped even if they are seen first.

https://htmlpreview.github.io/?https://github.com/nedbat/coverage-reports/blob/main/reports/20240314_1b19799edf/htmlcov/index.html
1b19799edf: master