94.623% - test: don't try to make pth files when invoked from pth #2011
When we're invoked by a .pth file, site.py hasn't fully run, so we try
to make a .pth file in the system Python instead of the virtualenv.
System Pythons are often not writable.

Also try to create this kind of not-writable arrangement in CI to find
problems like this in the future.

https://htmlpreview.github.io/?https://github.com/nedbat/coverage-reports/blob/main/reports/20250731_7a83ab0b9b/htmlcov/index.html
7a83ab0b9b: master