94.491% - test: find source directory relative to tests rather than `coverage` (#2028)
Use the path to the `tests/helpers.py` module itself rather than
the `coverage` package to determine what the source directory is.
The former is always imported from the source tree, while the latter
does not necessarily have to -- in particular, in Gentoo we are
installing it to a temporary directory to keep the source directory
clean of artifacts.

Signed-off-by: Michał Górny <mgorny@gentoo.org>

https://htmlpreview.github.io/?https://github.com/nedbat/coverage-reports/blob/main/reports/20250811_62434e79ef/htmlcov/index.html
62434e79ef: master